+++ changelog.md
@@ -6,6 +6,7 @@
 ## New features
 
 * Added a new utility function to color relationships `VisualizationGraph.color_relationships`
+* Added a new utility function to resize relationships `VisualizationGraph.resize_relationships`
 
 
 ## Bug fixes